+++ src/main.rs
@@ -1,8 +1,12 @@
 #![deny(missing_debug_implementations, rust_2018_idioms)]
 
+#[cfg(test)]
+mod tests;
+
 mod cargo;
 mod cli;
 mod config;
+mod cross_toml;
 mod docker;
 mod errors;
 mod extensions;
@@ -17,9 +21,10 @@ use std::path::PathBuf;
 use std::process::ExitStatus;
 
 use config::Config;
-use toml::{value::Table, Value};
+use serde::Deserialize;
 
 use self::cargo::{Root, Subcommand};
+use self::cross_toml::CrossToml;
 use self::errors::*;
 use self::rustc::{TargetList, VersionMetaExt};
 
@@ -87,6 +92,7 @@ impl Host {
         }
     }
 
+    /// Returns the [`Target`] as target triple string
     fn triple(&self) -> &str {
         match self {
             Host::X86_64AppleDarwin => "x86_64-apple-darwin",
@@ -116,7 +122,8 @@ impl<'a> From<&'a str> for Host {
     }
 }
 
-#[derive(Debug, Clone, PartialEq)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Deserialize)]
+#[serde(from = "&str")]
 pub enum Target {
     BuiltIn { triple: String },
     Custom { triple: String },
@@ -200,6 +207,20 @@ impl Target {
 
         !native && (self.is_linux() || self.is_windows() || self.is_bare_metal())
     }
+
+    fn needs_docker_privileged(&self) -> bool {
+        let arch_32bit = self.triple().starts_with("arm")
+            || self.triple().starts_with("i586")
+            || self.triple().starts_with("i686");
+
+        arch_32bit && self.is_android()
+    }
+}
+
+impl std::fmt::Display for Target {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_str(self.triple())
+    }
 }
 
 impl Target {
@@ -224,6 +245,13 @@ impl From<Host> for Target {
             Host::Aarch64UnknownLinuxMusl => Target::new_built_in("aarch64-unknown-linux-musl"),
             Host::Other(s) => Target::from(s.as_str(), &rustc::target_list(false).unwrap()),
         }
+    }
+}
+
+impl From<&str> for Target {
+    fn from(target_str: &str) -> Target {
+        let target_host: Host = target_str.into();
+        target_host.into()
     }
 }
 
@@ -253,14 +281,14 @@ fn run() -> Result<ExitStatus> {
         rustc_version::version_meta().wrap_err("couldn't fetch the `rustc` version")?;
     if let Some(root) = cargo::root()? {
         let host = version_meta.host();
-
-        if host.is_supported(args.target.as_ref()) {
-            let target = args
-                .target
-                .unwrap_or_else(|| Target::from(host.triple(), &target_list));
-            let toml = toml(&root)?;
-            let config = Config::new(toml);
-
+        let toml = toml(&root)?;
+        let config = Config::new(toml);
+        let target = args
+            .target
+            .or_else(|| config.target(&target_list))
+            .unwrap_or_else(|| Target::from(host.triple(), &target_list));
+        config.confusable_target(&target);
+        if host.is_supported(Some(&target)) {
             let mut sysroot = rustc::sysroot(&host, &target, verbose)?;
             let default_toolchain = sysroot
                 .file_name()
@@ -336,6 +364,12 @@ fn run() -> Result<ExitStatus> {
                     }
                 }
                 filtered_args
+            // Make sure --target is present
+            } else if !args.all.iter().any(|a| a.starts_with("--target")) {
+                let mut args_with_target = args.all.clone();
+                args_with_target.push("--target".to_string());
+                args_with_target.push(target.triple().to_string());
+                args_with_target
             } else {
                 args.all.clone()
             };
@@ -370,153 +404,9 @@ fn run() -> Result<ExitStatus> {
     cargo::run(&args.all, verbose)
 }
 
-/// Parsed `Cross.toml`
-#[derive(Debug)]
-pub struct Toml {
-    table: Table,
-}
-
-impl Toml {
-    /// Returns the `target.{}.image` part of `Cross.toml`
-    pub fn image(&self, target: &Target) -> Result<Option<String>> {
-        let triple = target.triple();
-
-        if let Some(value) = self
-            .table
-            .get("target")
-            .and_then(|t| t.get(triple))
-            .and_then(|t| t.get("image"))
-        {
-            Ok(Some(
-                value
-                    .as_str()
-                    .ok_or_else(|| eyre::eyre!("target.{triple}.image must be a string"))?
-                    .to_string(),
-            ))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Returns the `target.{}.runner` part of `Cross.toml`
-    pub fn runner(&self, target: &Target) -> Result<Option<String>> {
-        let triple = target.triple();
-
-        if let Some(value) = self
-            .table
-            .get("target")
-            .and_then(|t| t.get(triple))
-            .and_then(|t| t.get("runner"))
-        {
-            let value = value
-                .as_str()
-                .ok_or_else(|| eyre::eyre!("target.{triple}.runner must be a string"))?
-                .to_string();
-            Ok(Some(value))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Returns the `build.image` or the `target.{}.xargo` part of `Cross.toml`
-    pub fn xargo(&self, target: &Target) -> Result<(Option<bool>, Option<bool>)> {
-        let triple = target.triple();
-
-        if let Some(value) = self.table.get("build").and_then(|b| b.get("xargo")) {
-            return Ok((
-                Some(
-                    value
-                        .as_bool()
-                        .ok_or_else(|| eyre::eyre!("build.xargo must be a boolean"))?,
-                ),
-                None,
-            ));
-        }
-
-        if let Some(value) = self
-            .table
-            .get("target")
-            .and_then(|b| b.get(triple))
-            .and_then(|t| t.get("xargo"))
-        {
-            Ok((
-                None,
-                Some(
-                    value
-                        .as_bool()
-                        .ok_or_else(|| eyre::eyre!("target.{triple}.xargo must be a boolean"))?,
-                ),
-            ))
-        } else {
-            Ok((None, None))
-        }
-    }
-
-    /// Returns the list of environment variables to pass through for `build`,
-    pub fn env_passthrough_build(&self) -> Result<Vec<&str>> {
-        self.build_env("passthrough")
-    }
-
-    /// Returns the list of environment variables to pass through for `target`,
-    pub fn env_passthrough_target(&self, target: &Target) -> Result<Vec<&str>> {
-        self.target_env(target, "passthrough")
-    }
-
-    /// Returns the list of environment variables to pass through for `build`,
-    pub fn env_volumes_build(&self) -> Result<Vec<&str>> {
-        self.build_env("volumes")
-    }
-
-    /// Returns the list of environment variables to pass through for `target`,
-    pub fn env_volumes_target(&self, target: &Target) -> Result<Vec<&str>> {
-        self.target_env(target, "volumes")
-    }
-
-    fn target_env(&self, target: &Target, key: &str) -> Result<Vec<&str>> {
-        let triple = target.triple();
-
-        if let Some(&Value::Array(ref vec)) = self
-            .table
-            .get("target")
-            .and_then(|t| t.get(triple))
-            .and_then(|t| t.get("env"))
-            .and_then(|e| e.get(key))
-        {
-            vec.iter()
-                .map(|val| {
-                    val.as_str().ok_or_else(|| {
-                        eyre::eyre!("every target.{triple}.env.{key} element must be a string",)
-                    })
-                })
-                .collect()
-        } else {
-            Ok(Vec::new())
-        }
-    }
-
-    fn build_env(&self, key: &str) -> Result<Vec<&str>> {
-        if let Some(&Value::Array(ref vec)) = self
-            .table
-            .get("build")
-            .and_then(|b| b.get("env"))
-            .and_then(|e| e.get(key))
-        {
-            vec.iter()
-                .map(|val| {
-                    val.as_str().ok_or_else(|| {
-                        eyre::eyre!("every build.env.{key} element must be a string")
-                    })
-                })
-                .collect()
-        } else {
-            Ok(Vec::new())
-        }
-    }
-}
-
 /// Parses the `Cross.toml` at the root of the Cargo project or from the
 /// `CROSS_CONFIG` environment variable (if any exist in either location).
-fn toml(root: &Root) -> Result<Option<Toml>> {
+fn toml(root: &Root) -> Result<Option<CrossToml>> {
     let path = match env::var("CROSS_CONFIG") {
         Ok(var) => PathBuf::from(var),
         Err(_) => root.path().join("Cross.toml"),
@@ -525,22 +415,16 @@ fn toml(root: &Root) -> Result<Option<Toml>> {
     if path.exists() {
         let content = file::read(&path)
             .wrap_err_with(|| format!("could not read file `{}`", path.display()))?;
-        parse_toml(&content)
-            .wrap_err_with(|| format!("failed to parse file `{}` as TOML", path.display()))
+
+        let (config, _) = CrossToml::parse(&content)
+            .wrap_err_with(|| format!("failed to parse file `{}` as TOML", path.display()))?;
+
+        Ok(Some(config))
     } else {
-        // Let's check if there is a lower case version of the file
+        // Checks if there is a lowercase version of this file
         if root.path().join("cross.toml").exists() {
             eprintln!("There's a file named cross.toml, instead of Cross.toml. You may want to rename it, or it won't be considered.");
         }
         Ok(None)
     }
-}
-
-fn parse_toml(content: &str) -> Result<Option<Toml>> {
-    Ok(Some(crate::Toml {
-        table: match content.parse()? {
-            toml::Value::Table(table) => table,
-            _ => eyre::bail!("couldn't parse as TOML table"),
-        },
-    }))
 }